automatically resolve component types array types
This will ensure that component types of array types (e.g. `String` for `String[]`) are fully imported. Otherwise, things like annotations or class hierarchy will be missing for types that are not directly imported and only referenced through throws declarations.
As opposed to the other automatic resolutions I added this to the core resolve method in `ImportedClasses` because there are too many places where we also need to consider component types. Otherwise, this would never be robust for future changes (or at least always be a trip mine to overlook). To resolve the component type of any array type seems to be an intrinsic property of resolving any type, so this seems fair to me.
I thought about adding a shortcut to skip resolving further down the component type hierarchy if we encounter a type that has already been imported, but this would only help for repeated multi-dim arrays of the same component type but different dimensions (otherwise the cached type would be used anyway, and we don't resolve again). So, to me the benefit seems quite unclear, and we should probably wait until some bottleneck has shown before prematurely optimizing this at the cost of readability.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>